+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - 2.8.4
   - updated UI: highlight keywords for searching results
-  
+  - fixed coverage for static initialization blocks (#37)
 
 - 2.8.3
   - fixed error occurred in electron (#27)